# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bike Hub Api",
        "version": "0.1"
    },
    "paths": {
...
        "/api/threads": {
            "get": {
                "tags": [
                    "Threads"
                ],
                "summary": "Get all threads",
                "description": "Retrieve all threads with pagination.",
                "operationId": "55e1c4761d15d668ae62c513e5d40ac6",
                "responses": {
                    "200": {
                        "description": "List of threads retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/threads/{id}": {
            "put": {
                "tags": [
                    "Threads"
                ],
                "summary": "Update a thread",
                "description": "Update a thread by its ID.",
                "operationId": "b2c82963e6af11bdda96544891866d46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the thread to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Thread data to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "New Content"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Thread data updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Thread data updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Thread not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Thread not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Threads"
                ],
                "summary": "Delete a thread",
                "description": "Delete a thread by its ID.",
                "operationId": "f0bb440f46fd25adaa14f903454926a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the thread to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thread successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Thread successfully deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Thread not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Thread not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}